+++ scripts/merchants.js
@@ -1,0 +1,17 @@
+let merchants = {
+  floor1_merchant: {
+    name: "Mareen Halenfjor",
+    stock: [
+      {item: copy(consumables.inferior_healing_potion), price: 25, type: "consumable"},
+      {item: copy(consumables.inferior_mana_potion), price: 25, type: "consumable"},
+      {item: copy(consumables.lesser_healing_potion), price: 60, type: "consumable"},
+      {item: copy(weapons.broken_sword), price: 100, type: "weapon"},
+      {item: copy(weapons.broken_mace), price: 120, type: "weapon"},
+      {item: copy(weapons.battle_axe), price: 950, type: "weapon"},
+      {item: copy(weapons.long_sword), price: 365, type: "weapon"},
+      {item: copy(armors.rags), price: 50, type: "armor"},
+      {item: copy(armors.leather_armor), price: 280, type: "armor"},
+      {item: copy(armors.chainmail), price: 310, type: "armor"},
+    ]
+  }
+}